Add support for errors on inline snapshots
