+++ packages/jest-circus/src/legacy-code-todo-rewrite/jestAdapterInit.ts
@@ -32,6 +32,7 @@ import {
 } from '../state';
 import testCaseReportHandler from '../testCaseReportHandler';
 import {getTestID} from '../utils';
+import {jestAdapterEventTestDoneHandler} from '../eventsUtils';
 
 interface RuntimeGlobals extends Global.TestFrameworkGlobals {
   expect: JestExpect;
@@ -250,27 +251,8 @@ const eventHandler = async (event: Circus.Event) => {
       break;
     }
     case 'test_done': {
-      event.test.numPassingAsserts = jestExpect.getState().numPassingAsserts;
-      _addSuppressedErrors(event.test);
-      _addExpectedAssertionErrors(event.test);
+      jestAdapterEventTestDoneHandler(event);
       break;
     }
-  }
-};
-
-const _addExpectedAssertionErrors = (test: Circus.TestEntry) => {
-  const failures = jestExpect.extractExpectedAssertionsErrors();
-  const errors = failures.map(failure => failure.error);
-  test.errors = test.errors.concat(errors);
-};
-
-// Get suppressed errors from ``jest-matchers`` that weren't throw during
-// test execution and add them to the test result, potentially failing
-// a passing test.
-const _addSuppressedErrors = (test: Circus.TestEntry) => {
-  const {suppressedErrors} = jestExpect.getState();
-  jestExpect.setState({suppressedErrors: []});
-  if (suppressedErrors.length) {
-    test.errors = test.errors.concat(suppressedErrors);
   }
 };
